requirements.txt
azure-devops
GitPython
python-dateutil
pytz
trufflehog
pyyaml
requests
toml